+++ mem/include/toolib/mem/physmemusage-rss.h
@@ -43,16 +43,16 @@ void usage(double& virtual_mem_usage, double& physical_mem_usage)
         std::string O, itrealvalue, starttime;
     } dummy;
 
-    unsigned long vsize{};
-    long rss{};
+    double vsize{};
+    double rss{};
 
     stat_stream >> dummy.pid >> dummy.comm >> dummy.state >> dummy.ppid >> dummy.pgrp >> dummy.session >>
         dummy.tty_nr >> dummy.tpgid >> dummy.flags >> dummy.minflt >> dummy.cminflt >> dummy.majflt >> dummy.cmajflt >>
         dummy.utime >> dummy.stime >> dummy.cutime >> dummy.cstime >> dummy.priority >> dummy.nice >> dummy.O >>
         dummy.itrealvalue >> dummy.starttime >> vsize >> rss;
 
-    long page_size_kb = sysconf(_SC_PAGE_SIZE) / too::mem::KiB;
-    virtual_mem_usage = vsize / static_cast<double>(too::mem::KiB);
+    double page_size_kb = static_cast<double>(sysconf(_SC_PAGE_SIZE)) / too::mem::KiB;
+    virtual_mem_usage = vsize / too::mem::KiB;
     physical_mem_usage = rss * page_size_kb;
 #else
     too::ignore_arg(virtual_mem_usage);
